move GC etw enums to gcinterface.ee.h
