+++ src/gc/gcinterface.ee.h
@@ -5,6 +5,21 @@
 #ifndef _GCINTERFACE_EE_H_
 #define _GCINTERFACE_EE_H_
 
+enum EtwGCRootFlags
+{
+    kEtwGCRootFlagsPinning =            0x1,
+    kEtwGCRootFlagsWeakRef =            0x2,
+    kEtwGCRootFlagsInterior =           0x4,
+    kEtwGCRootFlagsRefCounted =         0x8,
+};
+
+enum EtwGCRootKind
+{
+    kEtwGCRootKindStack =               0,
+    kEtwGCRootKindFinalizer =           1,
+    kEtwGCRootKindHandle =              2,
+    kEtwGCRootKindOther =               3,
+};
 
 // This interface provides functions that the GC can use to fire events.
 // Events fired on this interface are split into two categories: "known"
