+++ src/inc/eventtracebase.h
@@ -38,22 +38,6 @@ void InitializeEventTracing();
 // These flags need to be defined either when FEATURE_EVENT_TRACE is enabled or the 
 // PROFILING_SUPPORTED is set, since they are used both by event tracing and profiling.
 
-enum EtwGCRootFlags
-{
-    kEtwGCRootFlagsPinning =            0x1,
-    kEtwGCRootFlagsWeakRef =            0x2,
-    kEtwGCRootFlagsInterior =           0x4,
-    kEtwGCRootFlagsRefCounted =         0x8,
-};
-
-enum EtwGCRootKind
-{
-    kEtwGCRootKindStack =               0,
-    kEtwGCRootKindFinalizer =           1,
-    kEtwGCRootKindHandle =              2,
-    kEtwGCRootKindOther =               3,
-};
-
 enum EtwTypeFlags
 {
     kEtwTypeFlagsDelegate =                         0x1,
